+++ frontend/lib/api/index.ts
@@ -1,3 +1,3 @@
-export { AdminAPI } from "./admin-client";
-export { PublicApi } from "./public-client";
-export { UserApiClient as UserApi } from "./user-client";
+export { AdminAPI } from "./client-admin";
+export { PublicApi } from "./client-public";
+export { UserApiClient as UserApi } from "./client-user";
